416: Move settings page to pages packages folder
