+++ packages/orchestrator-ui-components/src/components/WfoSettings/WfoSettings.tsx
@@ -2,29 +2,22 @@ import React from 'react';
 
 import { EuiHorizontalRule, EuiPageHeader, EuiSpacer } from '@elastic/eui';
 
-import { useEngineStatusMutation, useEngineStatusQuery } from '@/hooks';
 import { EngineStatus } from '@/types';
 
 import { WfoFlushSettings } from './WfoFlushSettings';
 import { WfoModifySettings } from './WfoModifySettings';
 import { WfoStatus } from './WfoStatus';
 
-export const WfoSettingsPage = () => {
-    const { data: engineStatus } = useEngineStatusQuery();
-    const { mutate, data: newEngineStatus } = useEngineStatusMutation();
-
-    const isRunning =
-        newEngineStatus?.global_status === EngineStatus.RUNNING ||
-        engineStatus?.global_status === EngineStatus.RUNNING;
-    const currentEngineStatus =
-        newEngineStatus?.global_status || engineStatus?.global_status;
-    const currentRunningProcesses =
-        newEngineStatus?.running_processes || engineStatus?.running_processes;
-
-    const changeEngineStatus = () => {
-        mutate({ global_lock: isRunning });
-    };
-
+interface WfoSettingsProps {
+    currentEngineStatus: EngineStatus | undefined;
+    currentRunningProcesses: number | undefined;
+    changeEngineStatus: () => void;
+}
+export const WfoSettings = ({
+    currentEngineStatus,
+    currentRunningProcesses,
+    changeEngineStatus,
+}: WfoSettingsProps) => {
     return (
         <>
             <EuiSpacer />
