+++ packages/orchestrator-ui-components/src/components/WfoSettings/index.ts
@@ -1,0 +1,1 @@
+export * from './WfoSettings';
