+++ packages/orchestrator-ui-components/src/components/index.ts
@@ -5,7 +5,7 @@ export * from './WfoKeyValueTable';
 export * from './WfoPageTemplate';
 export * from './WfoSearchBar';
 export * from './WfoSettingsModal';
-export * from './WfoSettingsPage';
+export * from '../pages/settings/WfoSettingsPage';
 export * from './WfoStartPage';
 export * from './WfoSubscription';
 export * from './WfoTable';
@@ -21,3 +21,4 @@ export * from './WfoProcessList';
 export * from './WfoTextAnchor';
 export * from './WfoJsonCodeBlock';
 export * from './WfoDiff';
+export * from './WfoSettings';
